requirements/requirements.txt
# Minimal requirements to run the application in production
# Conventions:
#  - Always pin requirements exactly (using ==). This prevents unexpected upgrades breaking the code
#  - Always include dependencies of packages. We install with --no-deps to prevent version clashes between dependencies
#  - Put an empty line and comment line above each requirement, with the dependencies directly underneath
Django==3.2.12
asgiref==3.5.0
sqlparse==0.4.2
pytz==2021.3

# Compatibility between Python 2 and 3
six==1.16.0

# REST API
djangorestframework==3.13.1

# Shell plus
django-extensions==3.1.5

# Use environment variables for configuration
django-environ==0.8.1

# Mysql database support
mysqlclient==2.1.0

# Enable login via LDAP
django-auth-ldap==4.0.0
python-ldap==3.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8

# Requests lib for making rest requests
requests==2.27.1
charset-normalizer==2.0.11

# Redis support
redis==4.4.0

# Store cache in Redis
django-redis-cache==3.0.0

# Uwsgi
uwsgi==2.0.20

# Deal with static files
whitenoise==5.3.0

# Django waffle, feature flipper
django-waffle==2.3.0

# Logging to Sentry
sentry-sdk==1.5.4
urllib3==1.26.8
certifi==2021.10.8
idna==3.3
chardet==4.0.0
more-itertools==8.6.0

# StatsD allows us to push monitoring data to a Graphite server
statsd==3.2.2
django-statsd-unleashed==1.1.1

# oauth2
oauthlib==3.2.0
requests-oauthlib==1.3.1
django-oauth-toolkit==1.5.0
cffi==1.15.0
cryptography==35.0.0
deprecated==1.2.13
jwcrypto==1.0
pycparser==2.21
wrapt==1.13.3

# Celery tasks
celery==4.4.0
kombu==4.6.7
billiard==3.6.1.0
amqp==2.5.2
vine==1.3.0
importlib_metadata==0.23
zipp==2.0.0
python-dateutil==2.8.1

# Use EnumFields in models
django-enumfields==2.1.1

# CORS headers
django-cors-headers==3.10.0

# Pybreaker
git+ssh://git@github.com/vikingco/pybreaker.git@57f10da36aae2eefc6d1c71952914238647d9a57#egg=pybreaker  #no-qa

# Convert html files to pdf
pyhtml2pdf==0.0.6
selenium==4.8.0
trio==0.22.0
trio-websocket==0.9.2
webdriver-manager==3.8.5
python-dotenv==0.21.1
tqdm==4.64.1
packaging==23.0
chromedriver-py==110.0.5481.30
